creator and constructor role
