+++ go.mod
@@ -6,6 +6,7 @@ require (
 	github.com/btcsuite/btcd v0.21.0-beta
 	github.com/btcsuite/btcutil v1.0.2
 	github.com/btcsuite/btcutil/psbt v1.0.2
+	github.com/go-ozzo/ozzo-validation/v4 v4.3.0
 	github.com/stretchr/testify v1.7.0
 	github.com/vulpemventures/fastsha256 v0.0.0-20160815193821-637e65642941
 	github.com/vulpemventures/go-secp256k1-zkp v1.1.3
