+++ psetv2/output.go
@@ -1,11 +1,11 @@
 package psetv2
 
 import (
-	"bytes"
-	"encoding/binary"
 	"errors"
 
-	"github.com/vulpemventures/go-elements/internal/bufferutil"
+	"github.com/vulpemventures/go-elements/elementsutil"
+
+	"github.com/vulpemventures/go-elements/transaction"
 )
 
 const (
@@ -84,356 +84,49 @@ type Output struct {
 	unknowns []keyPair
 }
 
-func (o *Output) serialize(s *bufferutil.Serializer) error {
-	outputKeyPairs, err := o.getKeyPairs()
-	if err != nil {
-		return err
-	}
+func psetOutputFromTxOutput(output transaction.TxOutput) (*Output, error) {
+	script := output.Script
 
-	for _, v := range outputKeyPairs {
-		if err := serializeKeyPair(v, s); err != nil {
-			return err
-		}
-	}
-
-	if err := s.WriteUint8(separator); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (o *Output) getKeyPairs() ([]keyPair, error) {
-	keyPairs := make([]keyPair, 0)
-
-	if o.redeemScript != nil {
-		redeemScriptKeyPair := keyPair{
-			key: key{
-				keyType: PsbtOutRedeemScript,
-				keyData: nil,
-			},
-			value: o.redeemScript,
-		}
-		keyPairs = append(keyPairs, redeemScriptKeyPair)
-	}
-
-	if o.witnessScript != nil {
-		witnessScriptKeyPair := keyPair{
-			key: key{
-				keyType: PsbtOutWitnessScript,
-				keyData: nil,
-			},
-			value: o.witnessScript,
-		}
-		keyPairs = append(keyPairs, witnessScriptKeyPair)
-	}
-
-	if o.bip32Derivation != nil {
-		for _, v := range o.bip32Derivation {
-			bip32DerivationPathKeyPair := keyPair{
-				key: key{
-					keyType: PsbtOutBip32Derivation,
-					keyData: v.PubKey,
-				},
-				value: SerializeBIP32Derivation(v.MasterKeyFingerprint, v.Bip32Path),
-			}
-			keyPairs = append(keyPairs, bip32DerivationPathKeyPair)
-		}
-	}
-
-	if o.outputScript != nil {
-		outputScriptKeyPair := keyPair{
-			key: key{
-				keyType: PsbtOutScript,
-				keyData: nil,
-			},
-			value: o.outputScript,
-		}
-		keyPairs = append(keyPairs, outputScriptKeyPair)
-	}
-
-	if o.outputValueCommitment != nil {
-		outputValueCommitmentKeyPair := keyPair{
-			key: key{
-				keyType: PsbtGlobalProprietary,
-				keyData: proprietaryKey(PsetElementsOutValueCommitment, nil),
-			},
-			value: o.outputValueCommitment,
-		}
-		keyPairs = append(keyPairs, outputValueCommitmentKeyPair)
-	}
-
-	if o.outputAmount != nil {
-		outputAmountBytes := make([]byte, 8)
-		binary.LittleEndian.PutUint64(outputAmountBytes, uint64(*o.outputAmount))
-		outputAmountKeyPair := keyPair{
-			key: key{
-				keyType: PsbtOutAmount,
-				keyData: nil,
-			},
-			value: outputAmountBytes,
-		}
-		keyPairs = append(keyPairs, outputAmountKeyPair)
-	}
-
-	if o.outputAssetCommitment != nil {
-		outputAssetCommitmentKeyPair := keyPair{
-			key: key{
-				keyType: PsbtGlobalProprietary,
-				keyData: proprietaryKey(PsetElementsOutAssetCommitment, nil),
-			},
-			value: o.outputAssetCommitment,
-		}
-		keyPairs = append(keyPairs, outputAssetCommitmentKeyPair)
-	}
-
-	if o.outputAsset != nil {
-		outputAssetKeyPair := keyPair{
-			key: key{
-				keyType: PsbtGlobalProprietary,
-				keyData: proprietaryKey(PsetElementsOutAsset, nil),
-			},
-			value: o.outputAsset,
-		}
-		keyPairs = append(keyPairs, outputAssetKeyPair)
-	}
-
-	if o.outputValueRangeproof != nil {
-		outputValueRangeproofKeyPair := keyPair{
-			key: key{
-				keyType: PsbtGlobalProprietary,
-				keyData: proprietaryKey(PsetElementsOutValueRangeproof, nil),
-			},
-			value: o.outputValueRangeproof,
-		}
-		keyPairs = append(keyPairs, outputValueRangeproofKeyPair)
-	}
-
-	if o.outputAssetSurjectionProof != nil {
-		outputAssetSurjectionProofKeyPair := keyPair{
-			key: key{
-				keyType: PsbtGlobalProprietary,
-				keyData: proprietaryKey(PsetElementsOutAssetSurjectionProof, nil),
-			},
-			value: o.outputAssetSurjectionProof,
-		}
-		keyPairs = append(keyPairs, outputAssetSurjectionProofKeyPair)
-	}
-
-	if o.outputBlindingPubkey != nil {
-		outputBlindingPubkeyKeyPair := keyPair{
-			key: key{
-				keyType: PsbtGlobalProprietary,
-				keyData: proprietaryKey(PsetElementsOutBlindingPubkey, nil),
-			},
-			value: o.outputBlindingPubkey,
-		}
-		keyPairs = append(keyPairs, outputBlindingPubkeyKeyPair)
-	}
-
-	if o.outputEcdhPubkey != nil {
-		outputEcdhPubkeyKeyPair := keyPair{
-			key: key{
-				keyType: PsbtGlobalProprietary,
-				keyData: proprietaryKey(PsetElementsOutEcdhPubkey, nil),
-			},
-			value: o.outputEcdhPubkey,
-		}
-		keyPairs = append(keyPairs, outputEcdhPubkeyKeyPair)
-	}
-
-	if o.outputBlinderIndex != nil {
-		outputBlinderIndexBytes := make([]byte, 4)
-		binary.LittleEndian.PutUint32(outputBlinderIndexBytes, *o.outputBlinderIndex)
-
-		outputBlinderIndexKeyPair := keyPair{
-			key: key{
-				keyType: PsbtGlobalProprietary,
-				keyData: proprietaryKey(PsetElementsOutBlinderIndex, nil),
-			},
-			value: outputBlinderIndexBytes,
-		}
-		keyPairs = append(keyPairs, outputBlinderIndexKeyPair)
-	}
-
-	if o.outputBlindValueProof != nil {
-		outputBlindValueProofKeyPair := keyPair{
-			key: key{
-				keyType: PsbtGlobalProprietary,
-				keyData: proprietaryKey(PsetElementsOutBlindValueProof, nil),
-			},
-			value: o.outputBlindValueProof,
-		}
-		keyPairs = append(keyPairs, outputBlindValueProofKeyPair)
-	}
-
-	if o.outputBlindAssetProof != nil {
-		outputBlindAssetProofKeyPair := keyPair{
-			key: key{
-				keyType: PsbtGlobalProprietary,
-				keyData: proprietaryKey(PsetElementsOutBlindAssetProof, nil),
-			},
-			value: o.outputBlindAssetProof,
-		}
-		keyPairs = append(keyPairs, outputBlindAssetProofKeyPair)
-	}
-
-	for _, v := range o.proprietaryData {
-		kp := keyPair{
-			key: key{
-				keyType: PsbtGlobalProprietary,
-				keyData: proprietaryKey(v.subtype, v.keyData),
-			},
-			value: v.value,
-		}
-		keyPairs = append(keyPairs, kp)
-	}
-
-	for _, v := range o.unknowns {
-		keyPairs = append(keyPairs, v)
-	}
-
-	return keyPairs, nil
-}
-
-func deserializeOutput(buf *bytes.Buffer) (*Output, error) {
-	output := &Output{}
-
-	kp := &keyPair{}
-
-	outputAssetFound := false
-	outputAmountFound := false
-	outputScriptFound := false
-
-	//read bytes and do the deserialization until separator is found at the
-	//end of global map
-	for {
-		if err := kp.deserialize(buf); err != nil {
-			if err == ErrNoMoreKeyPairs {
-				break
-			}
+	var outputAmount *int64
+	var outputCommitment []byte
+	if len(output.Value) == 9 && output.Value[0] == 1 {
+		o, err := elementsutil.ElementsToSatoshiValue(output.Value)
+		if err != nil {
 			return nil, err
 		}
-
-		switch kp.key.keyType {
-		case PsbtOutRedeemScript:
-			output.redeemScript = kp.value
-		case PsbtOutWitnessScript:
-			output.witnessScript = kp.value
-		case PsbtOutBip32Derivation:
-			if !validatePubkey(kp.key.keyData) {
-				return nil, ErrInvalidPsbtFormat
-			}
-			master, derivationPath, err := readBip32Derivation(kp.value)
-			if err != nil {
-				return nil, err
-			}
-
-			// Duplicate keys are not allowed
-			for _, x := range output.bip32Derivation {
-				if bytes.Equal(x.PubKey, kp.key.keyData) {
-					return nil, ErrDuplicatePubKeyInBip32DerPath
-				}
-			}
-
-			output.bip32Derivation = append(
-				output.bip32Derivation,
-				DerivationPathWithPubKey{
-					PubKey:               kp.key.keyData,
-					MasterKeyFingerprint: master,
-					Bip32Path:            derivationPath,
-				},
-			)
-		case PsbtOutAmount:
-			outputAmount := int64(binary.LittleEndian.Uint64(kp.value))
-			output.outputAmount = &outputAmount
-			outputAmountFound = true
-		case PsbtOutScript:
-			//TODO check if bellow validation is needed
-			//if len(kp.value) == 0 {
-			//	return nil, ErrInvalidOutputScriptLength
-			//}
-			output.outputScript = kp.value
-			outputScriptFound = true
-		case PsbtGlobalProprietary:
-			pd := &proprietaryData{}
-			if err := pd.proprietaryDataFromKeyPair(*kp); err != nil {
-				return nil, err
-			}
-
-			if bytes.Equal(pd.identifier, psetMagic[:len(psetMagic)-1]) {
-				switch pd.subtype {
-				case PsetElementsOutValueCommitment:
-					outValueCommitment := kp.value
-					if len(outValueCommitment) != 33 {
-						return nil, ErrInvalidValueCommitmentLength
-					}
-
-					outputAmountFound = true
-					output.outputValueCommitment = outValueCommitment
-				case PsetElementsOutAsset:
-					outputAsset := kp.value
-					if len(outputAsset) != 32 {
-						return nil, ErrInvalidOutAssetLength
-					}
-					output.outputAsset = outputAsset
-					outputAssetFound = true
-				case PsetElementsOutAssetCommitment:
-					outputAssetCommitment := kp.value
-					if len(outputAssetCommitment) != 33 {
-						return nil, ErrInvalidOutAssetCommitmentLength
-					}
-					outputAssetFound = true
-
-					output.outputAssetCommitment = outputAssetCommitment
-				case PsetElementsOutValueRangeproof:
-					output.outputValueRangeproof = kp.value
-				case PsetElementsOutAssetSurjectionProof:
-					output.outputAssetSurjectionProof = kp.value
-				case PsetElementsOutBlindingPubkey:
-					if !validatePubkey(kp.value) {
-						return nil, ErrInvalidPsbtFormat
-					}
-
-					output.outputBlindingPubkey = kp.value
-				case PsetElementsOutEcdhPubkey:
-					if !validatePubkey(kp.value) {
-						return nil, ErrInvalidPsbtFormat
-					}
-
-					output.outputEcdhPubkey = kp.value
-				case PsetElementsOutBlinderIndex:
-					outputBlinderIndex := binary.LittleEndian.Uint32(kp.value)
-					output.outputBlinderIndex = &outputBlinderIndex
-				case PsetElementsOutBlindValueProof:
-					output.outputBlindValueProof = kp.value
-				case PsetElementsOutBlindAssetProof:
-					output.outputBlindAssetProof = kp.value
-				default:
-					output.proprietaryData = append(output.proprietaryData, *pd)
-				}
-			}
-		default:
-			unknowns, err := deserializeUnknownKeyPairs(buf)
-			if err != nil {
-				return nil, err
-			}
-			output.unknowns = unknowns
+		ov := int64(o)
+		outputAmount = &ov
+	} else {
+		if len(output.Value) != 33 {
+			return nil, ErrInvalidAssetCommitmentLength
 		}
+
+		outputCommitment = output.Value
 	}
 
-	//validate mandatory fields
-	if !outputAssetFound {
-		return nil, ErrMissingOutAsset
+	var outputAsset []byte
+	var outputAssetCommitment []byte
+	if len(output.Asset) == 33 && output.Asset[0] == 1 {
+		outputAsset = output.Asset
+	} else {
+		if len(output.Value) != 33 {
+			return nil, ErrInvalidAssetCommitmentLength
+		}
+
+		outputAssetCommitment = output.Asset
 	}
 
-	if !outputAmountFound {
-		return nil, ErrMissingOutAmount
+	var outputBlindingPubkey []byte
+	if len(output.Nonce) == 33 && (output.Nonce[0] == 2 || output.Nonce[0] == 3) {
+		outputBlindingPubkey = output.Nonce
 	}
 
-	if !outputScriptFound {
-		return nil, ErrMissingScript
-	}
-
-	return output, nil
+	return &Output{
+		outputScript:          script,
+		outputAmount:          outputAmount,
+		outputValueCommitment: outputCommitment,
+		outputAsset:           outputAsset,
+		outputAssetCommitment: outputAssetCommitment,
+		outputBlindingPubkey:  outputBlindingPubkey,
+	}, nil
 }
